# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendor - Service",
        "description": "App Service documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/verify-pin": {
            "post": {
                "operationId": "b0d1e3a9168b382828ff70311b015720",
                "requestBody": {
                    "description": "request parameters to verify pin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PinValidationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/register-account": {
            "post": {
                "operationId": "a0965bc91eb2a0f86b8f67b45a789dbb",
                "requestBody": {
                    "description": "request parameters to register user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/number-category-validation/{number}": {
            "get": {
                "operationId": "0143634736aac612b342e3a415805f27",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Number",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                }
            }
        },
        "/api/existing-number/{number}": {
            "get": {
                "operationId": "c81f6ae466416e47e44d4debd66dd8e2",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Number",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PinValidationRequest": {
                "title": "Pin Validation Request",
                "description": "Pin Validation request",
                "required": [
                    "number",
                    "password"
                ],
                "properties": {
                    "number": {
                        "title": "number",
                        "description": "User Number",
                        "type": "string",
                        "example": "0557881327"
                    },
                    "password": {
                        "title": "password",
                        "description": "User Password or Pin",
                        "type": "string",
                        "example": "0000"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Register request",
                "required": [
                    "firstName",
                    "lastName",
                    "gender",
                    "mobileNumber"
                ],
                "properties": {
                    "firstName": {
                        "title": "firstName",
                        "description": "User first name",
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "User last name",
                        "type": "string",
                        "example": "Doe"
                    },
                    "gender": {
                        "title": "gender",
                        "description": "User's gender",
                        "type": "string",
                        "example": "m"
                    },
                    "mobileNumber": {
                        "title": "mobileNumber",
                        "description": "User's mobile number",
                        "type": "string",
                        "example": "0557001000"
                    },
                    "altMobileNumber": {
                        "title": "altMobileNumber",
                        "description": "User's alternative mobile number",
                        "type": "string",
                        "example": "0557001006"
                    },
                    "type": {
                        "title": "type",
                        "description": "Registration type",
                        "type": "string",
                        "example": "CUSTOMER / FIELD"
                    }
                },
                "type": "object"
            },
...
        }
    }
}